admin-portal/requirements.txt
fastapi
uvicorn
python-dotenv
motor
jinja2